Initial backend code outline check-in
Description:
This code is in an (obviously) unfinished state. :wink:

Some things to note are the field structures for the logic values
which use maps with string keys that would be concatenated
event-action names (this is subject to change). I've also taken
steps to try and keep the backend database logic as unit test
friendly as possible.

The logic code is now just direct interfaces (which could also
change) and I'm still exploring the possibility of implementing
multicasting within the backend.

Note: WIP check-in
Note: Unit tests have not been implemented