Add more pseudocode outlining processing logic
Description:
I've added the initial worker outlines; some of this will be
truncated versions of what we have in the legacy code (see the
notes for details). There is also a preliminary read() method
outlined on the database and the timer() method has been fleshed
out some more.

This is all still just to get a better picture of what we'll be
working with. :camera:

Note: WIP check-in
Note: No unit tests have been implemented