Expand primary settings check/update logic
Description:
This implements the bulk of the logic needed to identify which
responses need to be added to the repo. However, this does NOT
include the actual bootstrapping logic - this will likely be in
its own helper method/function for testability.

I've also included the server.go file since I updated some of the
pseudocode there.

Note: WIP check-in
Note: This check-in contains incomplete code
Note: A bare-bones unit test is passing