Revert settings TOML decoding to struct
Description:
Unfortunately, it doesn't seem like there's an easy option for using
the interface{} option in the TOML decoding easily :confused:. This
does mean that the internal issues struct will be (potentially) a
bit more difficult to read clearly but the front-facing arrangement
should still be easy.

I've included in-line comments for clarity.

Note: WIP check-in