Change database return structs for clarity
Description:
I've reverted the values in the integration struct to reflect the
non-pointers used in our old database because it's WAY easier that
way.

Also, I've removed the Integration field from the Options struct in
favor of keeping those two separate, but related, objects.

Note: WIP check-in
Note: Unit tests have been updated and are passing