Construct initial worker logic
Description:
This is the initial check-in for the actual worker.start() logic;
this is likely to change/get updated.

I've included a new work struct whose fields hold the values that
would be returned by the database query methods.

There is also a new dispatcher function (no unit test yet); there is
a new argument (vs the legacy version) so that we don't need to link
to the server (this value will change once we implement the
sync.Map type).

Note: WIP check-in
Note: There is (intentionally) missing processing logic currently
Note: Logging still needs to be implemented (for error scenarios)
Note: An initial unit test has been constructed for worker.start but
      I am not entirely sure it is correct/accurate (needs further
      evaluation)