Replace map with mutex struct
Description:
This is the planned transition of the server.repos field from the
concurrent-unsafe map to the struct with the embedded mutex; the
sync.Map had been an option buuut that was probably overkill so I
opted against it. :wink:

There are still probably a few areas where the mutex lock should be
envoked but that can be decided later as the logic is fleshed out.

Note: WIP check-in
Note: Unit tests have been updated and are passing