Build initial server Start method functionality
Description:
This is the method that is called when the server is launched; I've
made some changes from our legacy version (e.g. there is no listener
but that could change).

I've also approached this with the goal of making even this method
unit-testable (hence the openDatabase variable); note that the unit
test table here may be updated once I build the timer() method as it
doesn't actually do anything when called in Start() currently.

There is also a sync.WaitGroup here that I figured could be helpful
in booting up all of the repos from the database (this
implementation appears to be correct, but obviously it should be
reviewed for completeness).

Note: A bare-bonds unit test has been implemented for Start() and is
      passing
Note: All errors have been suppressed but will be reintroduced when
      logging is added to the codebase
Note: Response learn methods have also not yet been called in this
      check-in