Updating prod/test code for server assets
Description:
There were a few pretty important errors corrected in this commit in
the server stuff specifically (e.g. assigning to a nil internal map)
that's now fixed. I also opted to rename timer() to tick() as it's
more in-line with the standard library naming.

I did add an additional unit test to the server.Start() method but
for some reason it is continuing to fail - I'll look more at it
later. :grimacing: I've also introduced an interface for tick()
for unit testing but I'm not in love with it.

Note: Includes a breaking unit test