Restructure package imports for backend structs
Description:
I'm not entirely confident as to my choices here - I've opted to
include all of the necessary structs that hold objects either
immediately after retrieving them from the database or after some
intermediate processing has been done prior to passing them into
the Model methods. This SEEMS to smooth out potential hiccups in
importing packages and it does seem to "feel" logical but I'm not
sure it improves readability at all. I suspect that I can actually
include most of these simply in the process/ directory files.

I can circle back to this if it is an issue.

Note: Restructuring check-in
Note: Unit tests are passing but there is still a known race
      condition in the server_test.go file