Build logic for fetching new GitHub objects
Description:
This is the stuff necessary for pulling in new issues and pull
requests (and will likely be expanded to additional objects in the
future).

I opted to only go with single bulk insert methods for the database
since I sort of think the event objects added will be a bit
piecemeal; these have been added to the dataAccess interface.

I've also added in the necessary wrapper methods around the client
methods for fetching the objects. These still need unit tests.

Note: WIP check-in
Note: Unit tests have been updated and are passing but coverage
      is incomplete