Change parameters to accept new client functions
Description:
I made this change because unit testing otherwise was pretty
annoying with the reliance on a global variable which basically
just created race conditions in the tests. :angry: I did opt to
make a new newClient type that is the parameter rather than declare
anonymous functions for the parameters (stylistic choice only).

I am also considering looking for ways to make the "sub" worker
methods called within the "process" methods mockable but I have yet
to find a good implementation for this.

Note: WIP check-in
Note: This contains incomplete code