Re-implement repo initializer struct and logic
Description:
The value behind the initializer struct is clear to me now so I've
essentially added it back into the codebase - however, I did not
add it as a field onto the server struct and opted to keep it as an
in-memory value.

Because of this change, I've added in repo_init/_test files to hold
all relevant assets. There is only one unit test that is (partially)
implemented for the repoInit struct methods. The worker/_test files
have also been updated accordingly.

I've also added in the missing githubService interface methods and
made a few adjustments to where I was passing in client as a
parameter to make the code more unit-testable.

Note: WIP check-in
Note: This check-in includes in-code comments and incomplete code
Note: Unit tests are passing but not comprehensive