Add TOML fetching logic for addRepo
Description:
When a new repo is added, we need a way to pull down the
user-specified settings (in the .heupr.toml file). I've put it here
because I sort of expect the design of backend/ to read from the
database and process the contents as needed. :pensive:

I've updated all of the necessary interfaces and unit tests but
I have not yet added in scenarios in which TOML results are
1) retrieved via the client or 2) inserted into the database.

Note: WIP check-in
Note: This contains incomplete code and non-comprehensive unit tests
Note: Unit tests are passing
Note: Updates to addRepo have not been implemented yet