Adjust worker unit tests
Description:
Because we're utilizing several goroutines within the worker
methods, it's going to be difficult to actually test them without
bringing in a race condition for the tests. :disappointed: We can
circle back to this at some other time but for now, I'm just
commenting out the entirety of the worker_test.go contents.

Note: Unit test update
Note: Methods for worker will not have unit tests