Possible website redesign outline
Description:
One of the comments we've always gotten about the website is that it
does not clearly explain the service well enough - I'd like to do a
hand-built version that would be similar to our existing Hugo
version.

Right now, it's just in skeleton form with basic HTML structure and
a few pieces of CSS to get the idea. This would also be assuming
that we shift off of the control page and move our documentation to
a Markdown file on GitHub.

Note: WIP check-in