Cluster landing page buttons in separate element
Description:
When the screen size was changed, the buttons would not change in
the way I wanted them to so the are now in their own element. Also
several changes have been made to the responsiveness when they are
hovered over.

Note: WIP check-in
Note: Style sheet updates
Note: File loads correctly in browser