Flesh out landing page content
Description:
Got into a groove and wanted to bang out some of the (initial)
content that could go on the website. :pencil:

Right now, it's just text, but I'd like to add GIFs showing off the
features and some icons or still images to add color to the setup
text.

Note: WIP check-in
Note: Renders correctly from file in browser