Add "Team" section and images
Description:
I've added in a short section to identify the core builders and
also include links to some of our social media (we can update
this as needed).

I still need to make things adjustable for different view sizes
as well as update some of the links and possibly add smooth
scrolling between sections.

Note: index.html loads correctly when opened in browser