Build out initial backend processing module
Description:
Basic implementations of the "Learn" and "Predict" methods have
been added to the `backend/` module - these are specifically for
the "assignment" feature but are still officially incomplete.

Resources for modules have been included in order to include the
ES package. Additionally, a lot of mocking/wrapping has been
added in to make testing a little easier (in theory).

This check-in still leaves the application in an incomplete and
non-functional state.

Note: Full coverage summarization added to CI config
Note: Unit tests included and passing
Note: Closes #2
Note: Closes #13